Move ItemViewerControls to sibling component
